[Benja, Bere] adds authentication  logo
