+++ package.json
@@ -49,6 +49,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.9.6",
+    "sass": "^1.28.0",
     "typeorm": "^0.2.28"
   },
   "devDependencies": {
