+++ pages/login.ts
@@ -1,0 +1,1 @@
+export { Login as default } from 'src/steps/Login'
