+++ public/style.global.css
@@ -5,3 +5,6 @@ body {
   font: 14px 'Roboto', sans-serif;
   -webkit-font-smoothing: antialiased !important;
 }
+.content{
+  padding: 35px;
+}
